feat(client): HASHI-74 홈 페이지 퍼블리싱 코드리뷰 반영
