+++ apps/client/src/pages/home/components/HomeLogo.tsx
@@ -1,0 +1,13 @@
+import hashiLogoSrc from '@/shared/assets/logos/hashi-logo.svg'
+
+export const HomeLogo = () => {
+  return (
+    <img
+      alt="Hashi"
+      className="h-[23px] w-[73px]"
+      height={23}
+      src={hashiLogoSrc}
+      width={73}
+    />
+  )
+}
